Refactor : 케밥 아이콘 외 영역 누르면 팝오버 내려가는 기능 useRef사용하도록 수정
